// src/utils/loading/Loading.enum.ts
/**
 * Loading.enum.ts
 * Loading component enumerators.
 */
'use strict';

/**
 * Placeholder types.
 */
export enum PlaceholderType {
  EVENT_LIST = 'event-list',
  FILTER = 'filter',
  HEADER_IMAGE = 'header-image',
  LATEST_NEWS = 'latest-news',
  NEWS_LIST = 'news-list',
  PAGER = 'pager',
  SIDE_BANNER = 'side-banner',
  SIDE_MENU = 'side-menu',
  TAGS = 'tags',
  TEXT = 'text',
  TOP_BANNER = 'top-banner',
  TOP_MENU = 'top-menu'
}